BaseTools/Pccts: clean up antlr and dlg makefiles
(1) "-I" and "-D" options are for the preprocessor; move them to
    BUILD_CPPFLAGS. (This unifies BUILD_CPPFLAGS between both makefiles.)

(2) COTHER is never set, drop it from "antlr". (This unifies BUILD_CFLAGS
    between both makefiles, as COPT.)

(3) For linking "antlr" and "dlg", both BUILD_CFLAGS and BUILD_CPPFLAGS
    are useless, so drop BUILD_CFLAGS, and don't add BUILD_CPPFLAGS.

(4) For compiling C source files:

(4a) Move the "-c" mode selector to the front.

(4b) Expand both BUILD_CPPFLAGS and BUILD_CFLAGS, in this order. (This
     results in COPT being expanded last.)

(4c) Turn the source file operand into the last argument on the command
     line.

The only change in behavior from this patch is that the following options
disappear from the link-editing steps (due to (3)):

  -O -I. -I../support/set -I../h -DUSER_ZZSYN -DZZLEXBUFSIZE=65536

However these options made no difference for linking in the first place.

Cc: Liming Gao <liming.gao@intel.com>
Cc: Yonghong Zhu <yonghong.zhu@intel.com>
Ref: https://bugzilla.redhat.com/show_bug.cgi?id=1540244
Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Laszlo Ersek <lersek@redhat.com>