OvmfPkg/AcpiPlatformDxe: add [Un]RegisterAcpiTestSupport() skeletons
Introduce the PcdAcpiTestSupport Feature PCD to OvmfPkg. This PCD will
decide whether AcpiPlatformDxe populates the ACPI_TEST_SUPPORT structure
at the first Ready-To-Boot event.

(The Ready-To-Boot event is chosen because (a) it is bound to occur after
AcpiPlatformDxe installs the ACPI tables, and (b) it is guaranteed to
occur, namely when the boot manager launches the built-in UEFI shell.)

Introduce the RegisterAcpiTestSupport() and UnregisterAcpiTestSupport()
functions as skeletons, and call them dependent on the PCD, from the
driver's entry point function. Refactor the entry point function so the
registration can be rolled back on the failure path.

Cc: Anthony Perard <anthony.perard@citrix.com>
Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Cc: Drew Jones <drjones@redhat.com>
Cc: Igor Mammedov <imammedo@redhat.com>
Cc: Jordan Justen <jordan.l.justen@intel.com>
Cc: Julien Grall <julien.grall@linaro.org>
Cc: Philippe Mathieu-Daudé <philmd@redhat.com>
Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Laszlo Ersek <lersek@redhat.com>